removed references to old routing approach, migrated rest to new one
discuss: leave linkRoute as public or private? Acts as a quasai factory method for the Route object and adds it.

request method functions (get() post() put() delete()) are verbose to call nested in an addRoute(). move outside ot class so they don't need to be called by Route.get()?

some small additions to the staticjs and staticcss files, but they still need a way to know what folder to serve from. set asset route on the Dendro Object? use an env key? to decide